[FIX] website: allow use of custom search teamplate for the website search
With commit 123ba4ffccd6fc3895a18f3503690ba1593e3023 qweb.has_template get replaced with its owl equivalent. But right now the solution is not working and you can not use custom templates for search.

But with 10cdc570aa86c4c401ddb89368ce94af418cf071 owl version changed to version 2.2.9 which introduced the getRawTemplate method. So with this commit we replaced rawTemplates with getRawTemplate method. Which fixes the issue and allow you to use custom search templates for the website.